% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBuild_main.R
\name{Build_LineageMap_parallel}
\alias{Build_LineageMap_parallel}
\title{LineageMap main function: build tree with barcode, spatial coordinates and gene expression states, using parallelized subtree searches}
\usage{
Build_LineageMap_parallel(
  muts,
  meta,
  state_lineages,
  max_Iter = 200,
  threshold = 0.2,
  backbone_type = c("majority", "NJ_median", "NJ_mean"),
  outer_cores = 12,
  inner_cores = 5,
  lambda1 = 0.05,
  lambda2 = 0.1,
  alpha = 1
)
}
\arguments{
\item{muts}{lineage barcode matrix}

\item{meta}{spatial and state information of cells}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}

\item{max_Iter}{the maximum iterations for local search}

\item{threshold}{the threshold for building the similarity map}

\item{backbone_type}{the method used to reconstruct backbone}

\item{outer_cores}{the number of cores per outer loop}

\item{inner_cores}{the number of cores per inner loop (within FindBestTree)}

\item{lambda1}{weight for gene expression state likelihood}

\item{lambda2}{weight for spatial location likelihood}

\item{alpha}{hyperparameter for spatial likelihood}
}
\description{
LineageMap main function: build tree with barcode, spatial coordinates and gene expression states, using parallelized subtree searches
}
