% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Define_clones}
\alias{Define_clones}
\title{Getting clone ids from tree}
\usage{
Define_clones(tree, num_clones = 4)
}
\arguments{
\item{tree}{input lineage tree}

\item{num_clones}{number of desired clones}
}
\description{
Getting clone ids from tree
}
