% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DivideMut}
\alias{DivideMut}
\title{Neighbor Joining on unique lineage barcodes}
\usage{
DivideMut(X)
}
\arguments{
\item{X}{lineage barcode matrix}
}
\description{
Neighbor Joining on unique lineage barcodes
}
