% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_barcode.R
\name{drop_characters}
\alias{drop_characters}
\title{Simulate technical dropout uniformly across rows, cols, and globally}
\usage{
drop_characters(
  mat,
  drop_rate = 0.3,
  mode = c("global", "row", "col"),
  seed = NULL
)
}
\arguments{
\item{mat}{Unprocessed mutation barcodes}

\item{drop_rate}{Dropout rate, 0-1}

\item{mode}{Dropout mode globally, across rows or columns}

\item{seed}{random seed}
}
\description{
Simulate technical dropout uniformly across rows, cols, and globally
}
