% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_backbone.R
\name{DropoutDistMatrix}
\alias{DropoutDistMatrix}
\title{Calculate pairwise hamming distance matrix with dropouts}
\usage{
DropoutDistMatrix(X, dropout = "-", w_mut = 1, w_ref = 0.5)
}
\arguments{
\item{X}{barcode matrix}

\item{dropout}{dropout pattern}

\item{w_mut}{weight between 2 mutated barcodes}

\item{w_ref}{weight between an unmutated state and a mutated state}
}
\description{
Calculate pairwise hamming distance matrix with dropouts
}
