% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeBuild_main.R
\name{FindBestTree}
\alias{FindBestTree}
\title{Finding the best tree structure based on local search}
\usage{
FindBestTree(
  muts,
  meta,
  labels,
  state_lineages,
  newick_lookup = NULL,
  maxIter = 200,
  n_chains = 5,
  lambda1 = 0.05,
  lambda2 = 0.1,
  alpha = 1
)
}
\arguments{
\item{muts}{the lineage barcodes of cells}

\item{meta}{spatial coordinates and state information of cells}

\item{labels}{the labels of cells}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}

\item{newick_lookup}{lookup table for internal nodes}

\item{maxIter}{maximum number of iterations of local search}

\item{n_chains}{the number of chains to run}

\item{lambda1}{weight for gene expression state likelihood}

\item{lambda2}{weight for spatial location likelihood}

\item{alpha}{hyperparameter for spatial likelihood}
}
\description{
Finding the best tree structure based on local search
}
