% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_backbone.R
\name{GroupDistanceMatrix}
\alias{GroupDistanceMatrix}
\title{compute inter-group distances using pairwise distances}
\usage{
GroupDistanceMatrix(
  X,
  groups,
  dropout = "-",
  w_mut = 1,
  w_ref = 0.5,
  summary_fn = median
)
}
\arguments{
\item{X}{barcode matrix}

\item{groups}{subgroup ids of cells}

\item{dropout}{the dropout character}

\item{w_mut}{the weight between two mutations}

\item{w_ref}{weight between an unmutated state and a mutated state}

\item{summary_fn}{the function used to represent pairwise distance}
}
\description{
compute inter-group distances using pairwise distances
}
