% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{load_spatedsim}
\alias{load_spatedsim}
\title{LinRace main function(Dif): asymmetric division based Neighbor Joining}
\usage{
load_spatedsim(header, ncells, mr, rd, rm, run)
}
\arguments{
\item{muts}{lineage barcode matrix}

\item{states}{cell states of single cells}

\item{counts}{gene expression data of single cells}

\item{state_lineages}{the lineages that makes the state network from the root state to leaf states}

\item{max_Iter}{the maximum iterations for local search}

\item{lambda1}{hyperparameter for asymmetric division likelihood}

\item{lambda2}{hyperparameter for neighbor distance likelihood}
}
\description{
LinRace main function(Dif): asymmetric division based Neighbor Joining
}
