% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_backbone.R
\name{LouvainCluster}
\alias{LouvainCluster}
\title{Build similarity graph + Louvain}
\usage{
LouvainCluster(X, dropout = "-", threshold = 0.2, resolution = 1, dmat = NULL)
}
\arguments{
\item{X}{barcode matrix}

\item{dropout}{dropout pattern}

\item{threshold}{sparsity parameter for similarity graph}

\item{resolution}{clustering resolution for louvain}

\item{dmat}{pairwise distance matrix}
}
\description{
Build similarity graph + Louvain
}
