% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{NJ_from_barcode}
\alias{NJ_from_barcode}
\title{Neighbor Joining tree from lineage barcodes}
\usage{
NJ_from_barcode(X)
}
\arguments{
\item{X}{lineage barcode matrix}
}
\description{
Neighbor Joining tree from lineage barcodes
}
