% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_backbone.R
\name{SubgroupTree}
\alias{SubgroupTree}
\title{Build similarity graph + Louvain}
\usage{
SubgroupTree(X, groups, dmat = NULL)
}
\arguments{
\item{X}{barcode matrix}

\item{groups}{subgroup ids of cells}

\item{dmat}{pairwise distance matrix}
}
\description{
Build similarity graph + Louvain
}
