% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_backbone.R
\name{SubgroupTree_NJ}
\alias{SubgroupTree_NJ}
\title{Build the backbone tree using NJ on pairwise distances between subgroups}
\usage{
SubgroupTree_NJ(
  X,
  groups,
  dropout = "-",
  w_mut = 1,
  w_ref = 0.5,
  summary_fn = median
)
}
\arguments{
\item{X}{barcode matrix}

\item{groups}{subgroup ids of cells}
}
\description{
Build the backbone tree using NJ on pairwise distances between subgroups
}
