% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_Cal.R
\name{TreeLC2}
\alias{TreeLC2}
\title{tree local search function, subtree swapping}
\usage{
TreeLC2(tree, muts, loc, prob_internal = NULL)
}
\arguments{
\item{tree}{lineage tree object}

\item{muts}{lineage barcode data}

\item{loc}{spatial location data}

\item{prob_internal}{probability to sample every internal node}
}
\description{
tree local search function, subtree swapping
}
