% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_backbone.R
\name{WeightedDropoutHamming}
\alias{WeightedDropoutHamming}
\title{Calculate hamming distance with dropouts}
\usage{
WeightedDropoutHamming(a, b, dropout = "-", w_mut = 1, w_ref = 0.5)
}
\arguments{
\item{a}{barcode a}

\item{b}{barcode b}

\item{dropout}{dropout pattern}

\item{w_mut}{weight between 2 mutated barcodes}

\item{w_ref}{weight between an unmutated state and a mutated state}
}
\description{
Calculate hamming distance with dropouts
}
